# tables: optimize memory allocation in tables' hot path