--- a/pkg/sessionctx/variable/session.go
+++ b/pkg/sessionctx/variable/session.go
@@ -1632,8 +1632,10 @@ type TablesBufferPool struct {
 
 // UpdatePool is for UpdateRecord
 type UpdatePool struct {
-	ColIDs []int64
-	Row    []types.Datum
+	ColIDs     []int64
+	Row        []types.Datum
+	RowToCheck []types.Datum
+	ColSize    []ColSize
 }
 
 // GetOptimizerFixControlMap returns the specified value of the optimizer fix control.