--- a/pkg/table/contextimpl/table.go
+++ b/pkg/table/contextimpl/table.go
@@ -30,11 +30,25 @@ var _ context.AllocatorContext = &TableContextImpl{}
 type TableContextImpl struct {
 	sessionctx.Context
 	exprCtx exprctx.ExprContext
+
+	// TablesBuffer is a memory pool for table related memory allocation that aims to reuse memory
+	// and saves allocation
+	// The buffers are supposed to be used inside AddRecord/UpdateRecord/RemoveRecord.
+	// It's users duty to reset them before use.
+	TablesBuffer *context.TablesBuffer
 }
 
 // NewTableContextImpl creates a new TableContextImpl.
 func NewTableContextImpl(sctx sessionctx.Context, exprCtx exprctx.ExprContext) *TableContextImpl {
-	return &TableContextImpl{Context: sctx, exprCtx: exprCtx}
+	return &TableContextImpl{
+		Context: sctx,
+		exprCtx: exprCtx,
+		TablesBuffer: &context.TablesBuffer{
+			Add:    &context.AddRecordBuffer{},
+			Update: &context.UpdateRecordBuffer{},
+			Remove: &context.RemoveRecordBuffer{},
+		},
+	}
 }
 
 // TxnRecordTempTable record the temporary table to the current transaction.
@@ -50,4 +64,9 @@ func (ctx *TableContextImpl) GetExprCtx() exprctx.ExprContext {
 
 func (ctx *TableContextImpl) vars() *variable.SessionVars {
 	return ctx.Context.GetSessionVars()
+}
+
+// GetTablesBuffer implements the MutateContext interface.
+func (ctx *TableContextImpl) GetTablesBuffer() *context.TablesBuffer {
+	return ctx.TablesBuffer
 }