<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #53837

Problem Summary:

### What changed and how does it work?

Optimize memory allocations to save CPU time in the hot path.

Master:
```
go test -run=^$ -bench=Record -benchmem -benchtime=30s
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/pkg/table/tables
cpu: AMD Ryzen 5 3600 6-Core Processor
BenchmarkAddRecordInPipelinedDML-12       	    1050	  28595471 ns/op	      2860 ns/record	 7818797 B/op	  161352 allocs/op
BenchmarkRemoveRecordInPipelinedDML-12    	    2382	  12638043 ns/op	      1264 ns/record	 2848806 B/op	   90455 allocs/op
BenchmarkUpdateRecordInPipelinedDML-12    	    1844	  16815817 ns/op	      1682 ns/record	 8296254 B/op	  110608 allocs/op
PASS
ok  	github.com/pingcap/tidb/pkg/table/tables	261.149s
```

This PR:
```
go test -run=^$ -bench=Record -benchmem -benchtime=30s
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/pkg/table/tables
cpu: AMD Ryzen 5 3600 6-Core Processor
BenchmarkAddRecordInPipelinedDML-12       	    1569	  24222621 ns/op	      2422 ns/record	 5050033 B/op	  121106 allocs/op
BenchmarkRemoveRecordInPipelinedDML-12    	    3177	  10820754 ns/op	      1082 ns/record	 1736882 B/op	   70328 allocs/op
BenchmarkUpdateRecordInPipelinedDML-12    	    2612	  13296768 ns/op	      1330 ns/record	 3345944 B/op	   80497 allocs/op
PASS
ok  	github.com/pingcap/tidb/pkg/table/tables	288.480s
```

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test: existing tests and micro benchmark
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```