@cfzjywxk 2024-06-19
@ekexium 
Need approvements from
-  [pkg/lightning/OWNERS](https://github.com/pingcap/tidb/blob/master/pkg/lightning/OWNERS)
-  [pkg/sessionctx/variable/OWNERS](https://github.com/pingcap/tidb/blob/master/pkg/sessionctx/variable/OWNERS)