@ekexium 2024-06-19
@glorv FYI, this PR rewrites your previous optimization for lightning. cc @lance6716